Update: weather stats use compact integers over the wire instead of doubles
Rounds all weather values to integers and encodes them as the smallest
suitable type: temperature as int8 (signed, covers -128..127 °C), wind
speed / precipitation / cloud cover as uint8. Saves 7 bytes per field
per forecast hour. All layers updated in sync: backend struct packing,
frontend data handler vectors and stream reads, MainWeather slot and
signals, CurrentWeatherCard and WeatherForecastCard slots.
CLAUDE.md protocol table updated to reflect new wire formats.

+++ frontend/src/weather/datahandler/weatherdatahandler.cpp
@@ -21,10 +21,10 @@ void WeatherDataHandler::onMainForecastUpdate(const QByteArray &packet) {
     QIODevice *device = stream.device();
 
     QVector<quint8> times;
-    QVector<double> temperatures;
-    QVector<double> windspeeds;
-    QVector<double> precipitations;
-    QVector<double> cloudcovers;
+    QVector<qint8> temperatures;
+    QVector<quint8> windspeeds;
+    QVector<quint8> precipitations;
+    QVector<quint8> cloudcovers;
 
     while (true) {
         if (device->bytesAvailable() == 0) {
@@ -42,25 +42,25 @@ void WeatherDataHandler::onMainForecastUpdate(const QByteArray &packet) {
                 break;
 
             case 0x31:
-                double temperature;
+                qint8 temperature;
                 stream >> temperature;
                 temperatures.push_back(temperature);
                 break;
 
             case 0x32:
-                double windspeed;
+                quint8 windspeed;
                 stream >> windspeed;
                 windspeeds.push_back(windspeed);
                 break;
 
             case 0x33:
-                double precipitation;
+                quint8 precipitation;
                 stream >> precipitation;
                 precipitations.push_back(precipitation);
                 break;
 
             case 0x34:
-                double cloudcover;
+                quint8 cloudcover;
                 stream >> cloudcover;
                 cloudcovers.push_back(cloudcover);
                 break;
